+++ server/index.js
@@ -2,11 +2,13 @@ import express from 'express';
 import logger from 'morgan';
 import errorhandler from 'errorhandler';
 import cors from 'cors';
-import jwt from 'express-jwt';
+import expressJwt from 'express-jwt';
+import jwt from 'jsonwebtoken';
 import bodyParser from 'body-parser';
 import cookieParser from 'cookie-parser';
 
 import config from './config';
+import * as middleware from './middleware';
 import token from './resources/token';
 import user from './resources/user';
 
@@ -28,7 +30,7 @@ app.use(logger('dev'));
 app.use(errorhandler());
 
 const protect = () => {
-  return jwt({
+  return expressJwt({
     secret: config.secret,
     getToken: function (req) {
       return req.cookies.token;
@@ -39,15 +41,18 @@ const protect = () => {
 // set up API routing
 app.use('/api', protect(), router);
 router.post('/token', token.create);
-router.put('/token', token.update);
 router.delete('/token', token.delete);
-router.get('/user', user.read);
+router.get('/user', middleware.refreshToken, user.read);
 
 // Handle unauthorized errors gracefully.
 app.use(function (err, req, res, next) {
   if (err.name === 'UnauthorizedError') {
     res.status(401).send('Invalid Token');
   }
+});
+
+app.use(function (req, res, next) {
+
 });
 
 const server = app.listen(process.env.PORT || 3000, () => {
