+++ server/middleware/index.js
@@ -1,0 +1,1 @@
+export { default as refreshToken } from './refresh-token';
