+++ server/resources/token/index.js
@@ -9,30 +9,13 @@ export default {
     if (username === 'test' && password === 'test') {
       let token = jwt.sign({username: username}, config.secret, {
         issuer: 'redux-demo',
-        expiresInSeconds: 60,
+        expiresInSeconds: config.jwtExpiresInMinutes,
       });
 
       res.status(201).cookie('token', token).send();
     } else {
       res.status(401).send('Authentication failure.');
     }
-  },
-
-  update(req, res) {
-    const { cookies: { token: requestToken } } = req;
-    const decodedTokenPayload = jwt.decode(requestToken);
-    const { username } = decodedTokenPayload;
-
-    if (!username) {
-      return res.status(401).send('Invalid token');
-    }
-
-    const updatedToken = jwt.sign({username: username}, config.secret, {
-      issuer: 'redux-demo',
-      expiresInSeconds: 60,
-    });
-
-    res.status(204).cookie('token', updatedToken).send();
   },
 
   delete(req, res) {
